[PRMT-2684] removed logs which where putting a message in the logs
I also (temporarily) gave up to the attempt to make a single call to Parser class
from JmsConsumer (not 2 as we do now). Too many tests depends on the Jms message body
to be already formatted, and we haven't found a way to have the tests retrieving a message
from the file system to be correctly decoded by amqpMessage.decode.